requirements.txt
alembic==0.9.6
aniso8601==8.1.0
asgiref==3.3.1
Babel==2.5.1
bcrypt==3.2.0
blinker==1.4
certifi==2017.7.27.1
cffi==1.14.4
chardet==3.0.4
click==6.7
dnspython==2.0.0
dominate==2.3.1
elasticsearch==7.5.1
email-validator==1.1.2
Flask==1.0.2
Flask-Admin==1.5.7
Flask-Assets==2.0
Flask-Babel==0.11.2
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.11.0
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
gunicorn==19.7.1
idna==2.6
imgurpython==1.1.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
passlib==1.7.4
Pillow==8.1.0
pycparser==2.20
PyJWT==1.5.3
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
python-mimeparse==1.6.0
pytz==2017.2
requests==2.18.4
six==1.11.0
speaklater==1.3
SQLAlchemy==1.3.22
sqlparse==0.4.1
urllib3==1.22
visitor==0.1.3
webassets==2.0
Werkzeug==0.14.1
WTForms==2.1
psycopg2==2.8.6